requirements.txt
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
python-dotenv